+++ commands.go
@@ -129,6 +129,11 @@ func finishDeploy(i *Inspeqtor, args []string, resp io.Writer) {
 func currentStatus(i *Inspeqtor, args []string, resp io.Writer) {
 	io.WriteString(resp, fmt.Sprintf(
 		"%s %s, uptime: %s, pid: %d\n", Name, VERSION, time.Now().Sub(i.StartedAt).String(), os.Getpid()))
+
+	if i.silenced() {
+		io.WriteString(resp, fmt.Sprintf("Silenced until: %s\n", i.SilenceUntil.Format(time.RFC3339)))
+	}
+
 	io.WriteString(resp, "\n")
 
 	io.WriteString(resp, fmt.Sprintf("Host: %s\n", i.Host.Name()))