+++ Changes.md
@@ -11,6 +11,7 @@ See [Current Release Status](https://github.com/mperham/inspeqtor/milestones) he
 
 - Add tls_port option for SMTP servers. [exploid, #79]
 - Add gometalinter. [#83]
+- Show silenced until date in status output. [#101]
 
 ## 0.8.1-1